refactor: moved packages get command to its own file
