+++ lib/src/commands/commands.dart
@@ -1,5 +1,5 @@
 export 'create/commands/commands.dart';
 export 'create/create.dart';
-export 'packages.dart';
+export 'packages/packages.dart';
 export 'test/test.dart';
 export 'update.dart';
