+++ lib/src/commands/packages/commands/commands.dart
@@ -1,0 +1,1 @@
+export 'get.dart';
